+++ backend/src/main/java/ua/com/flowershop/service/FlowerService.java
@@ -38,7 +38,7 @@ public class FlowerService {
 
     public Page<FlowerFullProjection> findForAdmin(Long id, String flowerNamePart, String flowerOriginalNamePart, List<String> flowerTypeNames, String groupNamePart,
                                                    Integer sizeFrom, Integer sizeTo, Integer heightFrom, Integer heightTo,
-                                                   Integer popularityFrom, Integer popularityTo, String colorNamePart,
+                                                   Double popularityFrom, Double popularityTo, String colorNamePart,
                                                    LocalDateTime createdFrom, LocalDateTime createdTo,
                                                    Pageable pageRequest) {
         return flowerRepository.findForAdminProjectedByFilters(id, flowerNamePart, flowerOriginalNamePart, flowerTypeNames, groupNamePart, sizeFrom, sizeTo, heightFrom,
